+++ AzureIoTHub_Adapter/AzureIoTHub_Adapter.cs
@@ -36,7 +36,7 @@ namespace BH.Adapter.AzureIoTHub
     {
         public AzureIoTHubAdapter()
         {
-
+            AdapterIdName = "AzureIoTHub_Adapter";
         }
     }
 }
