+++ AzureIoTHub_oM/AzureIoTHub_oM.csproj
@@ -57,4 +57,8 @@
     <Compile Include="Temperature.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
+  <PropertyGroup>
+    <PostBuildEvent>xcopy "$(TargetDir)$(TargetFileName)"  "C:\ProgramData\BHoM\Assemblies" /Y
+</PostBuildEvent>
+  </PropertyGroup>
 </Project>
